requirements.txt
click==7.1.2
dnspython==2.0.0
enum34==1.1.10
eventlet==0.25.2
Flask==1.1.2
gevent==20.6.2
greenlet==0.4.16
gunicorn==20.0.4
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.19.0
opencv-python==4.3.0.36
paho-mqtt==1.5.0
RPi.GPIO-def==0.2
six==1.15.0
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
